+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 0.1.4 - Date: March 6, 2025
+
+- Refactored the dataclass scripts
+- Refactored the weather processor script into two separate scripts for current weather and forecast
+- Updated the README.md file
+
 ## Version 0.1.3 - Date: March 5, 2025
 
 - Initial release providing the current weather and forecast that connects with openweathermap API